+++ R/learndown-package.R
@@ -18,6 +18,7 @@
 #' @importFrom httr authenticate content GET parse_url status_code
 #' @importFrom utils compareVersion create.post packageVersion select.list
 #' @importFrom jsonlite base64_dec base64_enc fromJSON toJSON
+#' @importFrom PKI PKI.digest PKI.decrypt PKI.encrypt
 # The following block is used by usethis to automatically manage
 # roxygen namespace tags. Modify with care!
 ## usethis namespace: start